+++ dcmrt/libsrc/drtrcps.cc
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2014, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Source file for class DRTReferencedControlPointSequence
  *
- *  Generated automatically from DICOM PS 3.3-2017e
- *  File created on 2017-12-05 09:30:54
+ *  Generated automatically from DICOM PS 3.3-2007
+ *  File created on 2014-03-15 16:58:36
  *
  */
 
@@ -269,7 +269,7 @@ OFBool DRTReferencedControlPointSequence::isValid() const
 }
 
 
-size_t DRTReferencedControlPointSequence::getNumberOfItems() const
+unsigned long DRTReferencedControlPointSequence::getNumberOfItems() const
 {
     return SequenceOfItems.size();
 }
@@ -299,12 +299,12 @@ OFCondition DRTReferencedControlPointSequence::gotoNextItem()
 }
 
 
-OFCondition DRTReferencedControlPointSequence::gotoItem(const size_t num, OFListIterator(Item *) &iterator)
+OFCondition DRTReferencedControlPointSequence::gotoItem(const unsigned long num, OFListIterator(Item *) &iterator)
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        size_t idx = num + 1;
+        unsigned long idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -319,12 +319,12 @@ OFCondition DRTReferencedControlPointSequence::gotoItem(const size_t num, OFList
 }
 
 
-OFCondition DRTReferencedControlPointSequence::gotoItem(const size_t num, OFListConstIterator(Item *) &iterator) const
+OFCondition DRTReferencedControlPointSequence::gotoItem(const unsigned long num, OFListConstIterator(Item *) &iterator) const
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        size_t idx = num + 1;
+        unsigned long idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -339,7 +339,7 @@ OFCondition DRTReferencedControlPointSequence::gotoItem(const size_t num, OFList
 }
 
 
-OFCondition DRTReferencedControlPointSequence::gotoItem(const size_t num)
+OFCondition DRTReferencedControlPointSequence::gotoItem(const unsigned long num)
 {
     return gotoItem(num, CurrentItem);
 }
@@ -375,7 +375,7 @@ const DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence
 }
 
 
-OFCondition DRTReferencedControlPointSequence::getItem(const size_t num, Item *&item)
+OFCondition DRTReferencedControlPointSequence::getItem(const unsigned long num, Item *&item)
 {
     OFListIterator(Item *) iterator;
     OFCondition result = gotoItem(num, iterator);
@@ -385,7 +385,7 @@ OFCondition DRTReferencedControlPointSequence::getItem(const size_t num, Item *&
 }
 
 
-DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::getItem(const size_t num)
+DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::getItem(const unsigned long num)
 {
     OFListIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -395,7 +395,7 @@ DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::getI
 }
 
 
-const DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::getItem(const size_t num) const
+const DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::getItem(const unsigned long num) const
 {
     OFListConstIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -405,13 +405,13 @@ const DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence
 }
 
 
-DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::operator[](const size_t num)
+DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::operator[](const unsigned long num)
 {
     return getItem(num);
 }
 
 
-const DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::operator[](const size_t num) const
+const DRTReferencedControlPointSequence::Item &DRTReferencedControlPointSequence::operator[](const unsigned long num) const
 {
     return getItem(num);
 }
@@ -434,7 +434,7 @@ OFCondition DRTReferencedControlPointSequence::addItem(Item *&item)
 }
 
 
-OFCondition DRTReferencedControlPointSequence::insertItem(const size_t pos, Item *&item)
+OFCondition DRTReferencedControlPointSequence::insertItem(const unsigned long pos, Item *&item)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
@@ -457,7 +457,7 @@ OFCondition DRTReferencedControlPointSequence::insertItem(const size_t pos, Item
 }
 
 
-OFCondition DRTReferencedControlPointSequence::removeItem(const size_t pos)
+OFCondition DRTReferencedControlPointSequence::removeItem(const unsigned long pos)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
